refactor(transform,tools): #204 review — split oversized files touched by this PR
Pure moves, no behavior change:
- attribute_converter.go 585 -> 479 + attribute_value_extraction.go
  (extractValueFromEAVRecord + storage-mismatch error)
- attribute_converter_test.go 738 -> 361 + attribute_value_extraction_test.go
  + attribute_converter_list_test.go (carries the #204 list/empty-list tests)
- generate_attributes_test.go 1367 -> 240 + traverse/files/merge test files
