+++ internal/transform/attribute_converter_test.go
@@ -1,343 +1,14 @@
 package transform
 
 import (
-	"reflect"
-	"strconv"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/lychee-technology/forma/internal/model"
 
 	"github.com/google/uuid"
 	"github.com/lychee-technology/forma"
 )
-
-func TestExtractValueFromEAVRecord(t *testing.T) {
-	textVal := "hello"
-	numericVal := 42.0
-	uuidVal := uuid.New()
-	unixTime := time.Date(2024, 1, 2, 3, 4, 5, 0, time.UTC)
-	unixMillis := float64(unixTime.UnixMilli())
-
-	tests := []struct {
-		name      string
-		record    model.EAVRecord
-		valueType forma.ValueType
-		want      any
-		wantErr   string
-	}{
-		{
-			name:      "text nil returns nil",
-			record:    model.EAVRecord{},
-			valueType: forma.ValueTypeText,
-			want:      nil,
-		},
-		{
-			name: "text returns string",
-			record: model.EAVRecord{
-				ValueText: &textVal,
-			},
-			valueType: forma.ValueTypeText,
-			want:      "hello",
-		},
-		{
-			name: "smallint from numeric",
-			record: model.EAVRecord{
-				ValueNumeric: &numericVal,
-			},
-			valueType: forma.ValueTypeSmallInt,
-			want:      int16(42),
-		},
-		{
-			name: "integer from numeric",
-			record: model.EAVRecord{
-				ValueNumeric: &numericVal,
-			},
-			valueType: forma.ValueTypeInteger,
-			want:      int32(42),
-		},
-		{
-			name: "bigint from numeric",
-			record: model.EAVRecord{
-				ValueNumeric: &numericVal,
-			},
-			valueType: forma.ValueTypeBigInt,
-			want:      int64(42),
-		},
-		{
-			name: "numeric returns float64",
-			record: model.EAVRecord{
-				ValueNumeric: &numericVal,
-			},
-			valueType: forma.ValueTypeNumeric,
-			want:      numericVal,
-		},
-		{
-			name: "date from unix millis",
-			record: model.EAVRecord{
-				ValueNumeric: &unixMillis,
-			},
-			valueType: forma.ValueTypeDate,
-			want:      unixTime,
-		},
-		{
-			name: "datetime from unix millis",
-			record: model.EAVRecord{
-				ValueNumeric: &unixMillis,
-			},
-			valueType: forma.ValueTypeDateTime,
-			want:      unixTime,
-		},
-		{
-			name: "uuid from text",
-			record: model.EAVRecord{
-				ValueText: new(uuidVal.String()),
-			},
-			valueType: forma.ValueTypeUUID,
-			want:      uuidVal,
-		},
-		{
-			name: "uuid parse error",
-			record: model.EAVRecord{
-				ValueText: new("not-a-uuid"),
-			},
-			valueType: forma.ValueTypeUUID,
-			wantErr:   "parse uuid",
-		},
-		{
-			name: "bool true",
-			record: model.EAVRecord{
-				ValueNumeric: new(0.6),
-			},
-			valueType: forma.ValueTypeBool,
-			want:      true,
-		},
-		{
-			name: "bool false at threshold",
-			record: model.EAVRecord{
-				ValueNumeric: new(0.5),
-			},
-			valueType: forma.ValueTypeBool,
-			want:      false,
-		},
-		{
-			name: "unsupported uses text fallback",
-			record: model.EAVRecord{
-				ValueText: &textVal,
-			},
-			valueType: forma.ValueType("unknown"),
-			want:      "hello",
-		},
-		{
-			name: "unsupported uses numeric fallback",
-			record: model.EAVRecord{
-				ValueNumeric: &numericVal,
-			},
-			valueType: forma.ValueType("unknown"),
-			want:      numericVal,
-		},
-		{
-			name:      "unsupported nil returns nil",
-			record:    model.EAVRecord{},
-			valueType: forma.ValueType("unknown"),
-			want:      nil,
-		},
-		{
-			name: "numeric nil returns nil",
-			record: model.EAVRecord{
-				ValueNumeric: nil,
-			},
-			valueType: forma.ValueTypeNumeric,
-			want:      nil,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := extractValueFromEAVRecord(tt.record, tt.valueType)
-			if tt.wantErr != "" {
-				if err == nil {
-					t.Fatalf("expected error %q, got nil", tt.wantErr)
-				}
-				if !strings.Contains(err.Error(), tt.wantErr) {
-					t.Fatalf("expected error %q, got %q", tt.wantErr, err.Error())
-				}
-				return
-			}
-			if err != nil {
-				t.Fatalf("unexpected error: %v", err)
-			}
-			switch want := tt.want.(type) {
-			case time.Time:
-				gotTime, ok := got.(time.Time)
-				if !ok {
-					t.Fatalf("expected time.Time, got %T", got)
-				}
-				if !gotTime.Equal(want) {
-					t.Fatalf("expected time %v, got %v", want, gotTime)
-				}
-			case uuid.UUID:
-				gotUUID, ok := got.(uuid.UUID)
-				if !ok {
-					t.Fatalf("expected uuid.UUID, got %T", got)
-				}
-				if gotUUID != want {
-					t.Fatalf("expected uuid %v, got %v", want, gotUUID)
-				}
-			case nil:
-				if got != nil {
-					t.Fatalf("expected nil, got %v", got)
-				}
-			default:
-				if !reflect.DeepEqual(got, want) {
-					t.Fatalf("expected %v (%T), got %v (%T)", want, want, got, got)
-				}
-			}
-		})
-	}
-}
-
-// newListStubRegistry builds a registry whose schema carries two list
-// attributes: text items (default) and integer items.
-func newListStubRegistry() forma.SchemaRegistry {
-	return &stubSchemaRegistry{
-		schemaID:   100,
-		schemaName: "test",
-		cache: forma.SchemaAttributeCache{
-			"tags": {AttributeName: "tags", AttributeID: 18, ValueType: forma.ValueTypeList},
-			"nums": {AttributeName: "nums", AttributeID: 19, ValueType: forma.ValueTypeList, ItemsType: forma.ValueTypeInteger},
-		},
-	}
-}
-
-func TestAttributeConverterFromEAVRecords_ListElementsTypedByItemsType(t *testing.T) {
-	registry := newListStubRegistry()
-	converter := NewAttributeConverter(registry)
-	rowID := uuid.New()
-
-	text := func(s string) *string { return &s }
-	num := func(f float64) *float64 { return &f }
-	records := []model.EAVRecord{
-		{SchemaID: 100, RowID: rowID, AttrID: 18, ArrayIndices: "0", ValueText: text("alpha")},
-		{SchemaID: 100, RowID: rowID, AttrID: 18, ArrayIndices: "1", ValueText: text("beta")},
-		{SchemaID: 100, RowID: rowID, AttrID: 19, ArrayIndices: "0", ValueNumeric: num(7)},
-	}
-
-	attrs, err := converter.FromEAVRecords(records)
-	if err != nil {
-		t.Fatalf("FromEAVRecords: %v", err)
-	}
-	if len(attrs) != 3 {
-		t.Fatalf("got %d attributes, want 3", len(attrs))
-	}
-	for _, attr := range attrs {
-		switch {
-		case attr.AttrID == 18:
-			if attr.ValueType != forma.ValueTypeText {
-				t.Errorf("tags element ValueType = %q, want text (items type)", attr.ValueType)
-			}
-			if _, ok := attr.Value.(string); !ok {
-				t.Errorf("tags element value = %T(%v), want string", attr.Value, attr.Value)
-			}
-		case attr.AttrID == 19:
-			if attr.ValueType != forma.ValueTypeInteger {
-				t.Errorf("nums element ValueType = %q, want integer (items type)", attr.ValueType)
-			}
-			if _, ok := attr.Value.(int32); !ok {
-				t.Errorf("nums element value = %T(%v), want int32", attr.Value, attr.Value)
-			}
-		}
-	}
-}
-
-func TestPersistentRecord_ListRoundTrip(t *testing.T) {
-	ctx := t.Context()
-	registry := newListStubRegistry()
-	prt := NewPersistentRecordTransformer(registry)
-	rowID := uuid.Must(uuid.NewV7())
-
-	record, err := prt.ToPersistentRecord(ctx, 100, rowID, map[string]any{
-		"tags": []any{"alpha", "beta", "gamma"},
-		"nums": []any{1, 2},
-	})
-	if err != nil {
-		t.Fatalf("ToPersistentRecord: %v", err)
-	}
-	if len(record.OtherAttributes) != 5 {
-		t.Fatalf("got %d EAV records, want 5 (3 tags + 2 nums): %+v", len(record.OtherAttributes), record.OtherAttributes)
-	}
-	byKey := make(map[string]model.EAVRecord)
-	for _, r := range record.OtherAttributes {
-		byKey[strconv.Itoa(int(r.AttrID))+"|"+r.ArrayIndices] = r
-	}
-	for i, want := range []string{"alpha", "beta", "gamma"} {
-		r, ok := byKey["18|"+strconv.Itoa(i)]
-		if !ok || r.ValueText == nil || *r.ValueText != want {
-			t.Errorf("tags[%d]: got %+v, want value_text %q", i, r, want)
-		}
-	}
-	for i, want := range []float64{1, 2} {
-		r, ok := byKey["19|"+strconv.Itoa(i)]
-		if !ok || r.ValueNumeric == nil || *r.ValueNumeric != want {
-			t.Errorf("nums[%d]: got %+v, want value_numeric %v", i, r, want)
-		}
-	}
-
-	// Round trip back to JSON rebuilds ordered arrays.
-	rebuilt, err := prt.FromPersistentRecord(ctx, record)
-	if err != nil {
-		t.Fatalf("FromPersistentRecord: %v", err)
-	}
-	if !reflect.DeepEqual(rebuilt["tags"], []any{"alpha", "beta", "gamma"}) {
-		t.Errorf("rebuilt tags = %#v, want [alpha beta gamma]", rebuilt["tags"])
-	}
-	nums, ok := rebuilt["nums"].([]any)
-	if !ok || len(nums) != 2 {
-		t.Fatalf("rebuilt nums = %#v, want 2-element array", rebuilt["nums"])
-	}
-}
-
-func TestExtractValueFromEAVRecord_ReturnsErrorOnStorageTypeMismatch(t *testing.T) {
-	textVal := "123"
-	_, err := extractValueFromEAVRecord(model.EAVRecord{ValueText: &textVal}, forma.ValueTypeNumeric)
-	if err == nil {
-		t.Fatal("expected storage/type mismatch error")
-	}
-	if !strings.Contains(err.Error(), "storage type mismatch") || !strings.Contains(err.Error(), "value_text should not be populated") {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
-
-func TestExtractValueFromEAVRecord_BothColumnsPopulatedReturnsError(t *testing.T) {
-	textVal := "text"
-	numericVal := 123.0
-
-	tests := []struct {
-		name      string
-		valueType forma.ValueType
-		wantErr   string
-	}{
-		{name: "text type", valueType: forma.ValueTypeText, wantErr: "value_numeric should not be populated"},
-		{name: "numeric type", valueType: forma.ValueTypeNumeric, wantErr: "value_text should not be populated"},
-		{name: "uuid type", valueType: forma.ValueTypeUUID, wantErr: "value_numeric should not be populated"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := extractValueFromEAVRecord(model.EAVRecord{
-				ValueText:    &textVal,
-				ValueNumeric: &numericVal,
-			}, tt.valueType)
-			if err == nil {
-				t.Fatal("expected storage/type mismatch error")
-			}
-			if !strings.Contains(err.Error(), "storage type mismatch") || !strings.Contains(err.Error(), tt.wantErr) {
-				t.Fatalf("unexpected error: %v", err)
-			}
-		})
-	}
-}
 
 func TestToFloat64ForEAV(t *testing.T) {
 	tests := []struct {
